refactor: separate disparate concerns in main
The project's driver file currently defines a stand-alone class, some
helper functions, then provides a means of running them. Separate this
class, the JavaProgram class, out into its own file and move the driver
functionality, including some helper functions, out to their own
file.
