+++ main.py
@@ -1,10 +1,4 @@
-"""
-main.py
-
-By Benjamin Carter, Dylan Johnson, and Hunter Jenkins
-
-This program constitutes the main of the CoreEngine
-"""
+"""TODO."""
 
 import glob
 import os
@@ -16,8 +10,8 @@ from src.AI_Taxonomy import AICachedClassifier, load_data
 from src.DatabaseManager import DatabaseManager
 from src import github_pull
 from src.generateAST import generateAST
-from src.symbolTable import SymbolTable
-from src import tokens as tokenExtract
+from src.java_ast import JavaProgram
+
 from src import store_result
 from src import database_init
 
@@ -27,196 +21,55 @@ YELLOW_COLOR = "\033[1m\033[38;5;11m"
 RESET_COLOR = "\033[0m"
 
 
-class JavaProgram:
-    """Class represents a Java AST Tree."""
+def main():
+    """TODO."""
 
-    def __init__(self, java_ast: dict):
-        """Set up the class. Requires AST.
+    api_listing_file = "./data/domain_labels.json"
+    sub_domain_listing_file = "./data/subdomain_labels.json"
 
-        Args:
-            java_ast (dict): AST
+    # Setup!
+    if not os.path.isdir("output"):
+        os.makedirs("output")
+
+    os.chdir("output")
+    if not os.path.isdir("downloadedFiles"):
+        os.makedirs("downloadedFiles")
+    os.chdir("../")
+
+    def setup_db():
         """
-        self.ast = java_ast
-
-        self.classes = None
-        self.plain_classes = None
-        self.methods = None
-        self.symbols = None
-        self.complete_table = None
-        self.functions = None
-
-    def get_classes(self):
-        """Get classes in program
-
-        Returns:
-            set[str]: Set of all classes in program
-        """
-        if self.plain_classes is not None:
-            return self.plain_classes
-
-        self.plain_classes = set()
-
-        class_options = self.get_class_options()
-        for class_name in class_options.items():
-            node = class_options[class_name]
-            if node == 0:
-                continue
-            for x in node:
-                self.plain_classes.add(x)
-        return self.plain_classes
-
-    def get_class_options(self):
-        """
-        Takes the classnames from the file, matches it to imports, and
-        returns all the full-name class names.
-
-        Returns:
-            list: Data structure with the tokens matched to the full
-            class name based off of imports.
+        Put instructions here that you want run on first initialization
+        (after folders are made)
         """
 
-        if self.classes is not None:
-            return self.classes
+        # Clear download directory
+        for file in glob.glob("output/downloadedFiles/*"):
+            os.remove(file)
 
-        ast = self.ast
+        database_init.populate_db_with_mining_data()
+        database_init.setup_caches()
 
-        tokens = set(tokenExtract.pullToken(ast))
-        imports = set(tokenExtract.pullImport(ast))
+    database_init.start(setup_db)
 
-        result = {}
-        import_items = {}
-        for i in imports:
-            params = i.split(".")
-            name = params[-1]
-            if name in import_items:
-                import_items[name].append(i)
-            else:
-                import_items[name] = [i]
+    # Create database connection
+    db = DatabaseManager()
 
-        for token in tokens:
-            # check to see if it is in the imports.
-            # if not, check lang
-            # if not, fail
-            if token in import_items:
-                result[token] = import_items[token]
-            else:
-                result[token] = 0
+    api_domain_listing = load_data(api_listing_file)
+    sub_domain_listing = load_data(sub_domain_listing_file)
 
-        self.classes = result
-        return result
+    classifier = AICachedClassifier(api_domain_listing, sub_domain_listing, db)
 
-    def populate_symbol_table(self):
-        """Pre-generate symbols and methods"""
+    print("\nProcessing files...")
+    process_files(classifier, db)
 
-        pgrm_tables = SymbolTable(self.ast)
+    db.save()
+    db.close()
 
-        self.symbols = (
-            pgrm_tables.findSymbols()
-        )  # gets all classes with variable names.
-
-        self.methods = pgrm_tables.getMethods()  # gets all methods from variable name.
-
-    def get_complete_symbol_table(self):
-        """Match methods and symbols and class names into a combined data structure.
-
-        Returns:
-            dict: Data Structure of compiled symbols, methods, and classes.
-        """
-
-        if self.classes is None:
-            self.get_class_options()
-
-        if self.symbols is None or self.methods is None:
-            self.populate_symbol_table()
-
-        if self.complete_table is not None:
-            return self.complete_table
-
-        out = {}
-        for class_name in self.classes:
-            out[class_name] = {"full": self.classes[class_name], "varlist": []}
-
-            for num, variable in enumerate(self.symbols):
-                if variable["class"] != class_name:
-                    continue
-                variable_name = variable["name"]
-
-                for vr in self.symbols[num + 1 :]:
-                    # see if there are any more!
-                    if vr["class"] != class_name:
-                        continue
-                    # if(variable["line"] > lineNumber):
-                    break  # CHECK! We may need to move this back.
-
-                method_out = []
-                for method in self.methods:
-                    # check to see if it is referencing the same variable
-                    if method["name"] != variable_name:
-                        continue
-
-                    # TODO: instead of comparing line numbers, compare depth in AST tree.
-                    # print(className, method["name"], lineNumber, nextLine, method["line"])
-                    # if(method["line"] < lineNumber):
-                    #     continue
-                    # if(nextLine != -1 and method["line"] > nextLine):
-                    #     continue
-                    # print("Append!")
-                    method_out.append(method)
-
-                out[class_name]["varlist"].append(
-                    {"variable": variable, "methods": method_out}
-                )
-
-        self.complete_table = out
-        return out
-
-    def get_functions(self):
-        """Return all the functions used in the program
-
-        Returns:
-            dict[str : list]: function_name : [line numbers]
-        """
-        if self.functions is not None:
-            return self.functions
-
-        if self.complete_table is None:
-            self.get_complete_symbol_table()
-
-        functions = {}
-        for class_name in self.complete_table:
-            data = self.complete_table[class_name]
-
-            methods_general = []
-            for varl in data["varlist"]:
-                for m in varl["methods"]:
-                    methods_general.append(m)
-
-            if data["full"] != 0:
-                for f_n in data["full"]:
-                    for m_n in methods_general:
-                        if f"{f_n}::{m_n['method']}" in functions:
-                            functions[f"{f_n}::{m_n['method']}"].append(m_n["line"])
-                        else:
-                            functions[f"{f_n}::{m_n['method']}"] = [m_n["line"]]
-            else:
-                for m_n in methods_general:
-                    if f"Unknown::{m_n['method']}" in functions:
-                        functions[f"Unknown::{m_n['method']}"].append(m_n["line"])
-                    else:
-                        functions[f"Unknown::{m_n['method']}"] = [m_n["line"]]
-
-        self.functions = functions
-        return functions
-
-    # Function to extract and print class names and methods
-    def extract_classes_and_methods(self) -> tuple[set[str], dict]:
-        """Function to extract and print class names and methods
-
-        Returns:
-            set: classes
-            dict: functions
-        """
-        return self.get_classes(), self.get_functions()
+    # EXPORT! Requires DB to be closed.
+    print("\nExporting....")
+    store_result.sqlite_to_csv(
+        "./output/main.db", "outputTable", "./output/core_engine_output.csv"
+    )
 
 
 def process_files(ai: AICachedClassifier, database: DatabaseManager):
@@ -329,50 +182,4 @@ def process_files(ai: AICachedClassifier, database: DatabaseManager):
 
 
 if __name__ == "__main__":
-    # Setup!
-    if not os.path.isdir("output"):
-        os.makedirs("output")
-
-    os.chdir("output")
-    if not os.path.isdir("downloadedFiles"):
-        os.makedirs("downloadedFiles")
-    os.chdir("../")
-
-    def setup_db():
-        """
-        Put instructions here that you want run on first initialization
-        (after folders are made)
-        """
-
-        # Clear download directory
-        for file in glob.glob("output/downloadedFiles/*"):
-            os.remove(file)
-
-        database_init.populate_db_with_mining_data()
-        database_init.setup_caches()
-
-    database_init.start(setup_db)
-
-    # Create database connection
-    db = DatabaseManager()
-    # -----------------------
-    API_LISTING_FILE = "./data/domain_labels.json"
-    SUB_DOMAIN_LISTING_FILE = "./data/subdomain_labels.json"
-    # -----------------------
-
-    api_domain_listing = load_data(API_LISTING_FILE)
-    sub_domain_listing = load_data(SUB_DOMAIN_LISTING_FILE)
-
-    classifier = AICachedClassifier(api_domain_listing, sub_domain_listing, db)
-
-    print("\nProcessing files...")
-    process_files(classifier, db)
-
-    db.save()
-    db.close()
-
-    # EXPORT! Requires DB to be closed.
-    print("\nExporting....")
-    store_result.sqlite_to_csv(
-        "./output/main.db", "outputTable", "./output/core_engine_output.csv"
-    )
+    main()
